Add staticcheck, govet, errcheck, gosec, gocyclo, misspell, unconvert, unparam linters
Fix all new lint errors:
- Replace crypto/md5 (weak primitive) with hash/fnv in temp dir name generation
  in workflow_base.go and repo.go
- Remove redundant int64() cast on forkProject.ID (already int64) in
  composer_patches_1.go
- Exclude gosec G204/G702 (subprocess taint) from test files — the
  TestHelperProcess pattern is standard Go practice, not a real risk
- Add nolint:gosec to rector_test.go subprocess lines for same reason

gosimple was omitted — it was merged into staticcheck in golangci-lint v2.

+++ internal/addon/composer_patches_1.go
@@ -333,7 +333,7 @@ func (h *ComposerPatches1) processSinglePatch(ctx context.Context, path string, 
 	}
 	h.logger.Debug("fetched fork project", zap.Any("project", forkProject))
 
-	mergeRequests, err := h.fetchForkMergeRequests(issue.Project.MaschineName, int64(forkProject.ID))
+	mergeRequests, err := h.fetchForkMergeRequests(issue.Project.MaschineName, forkProject.ID)
 	if err != nil {
 		return
 	}
